Soyuz satellite placement begun
